requirements.txt
praw>=7.7.1
textblob>=0.17.1
psycopg2-binary>=2.9.9
pydantic>=2.5.2
pydantic-settings>=2.1.0
python-dotenv>=1.0.0 